+++ Service/TextBuilding/PersonalLogTextBuilderBase.cs
@@ -1,0 +1,39 @@
+using System.Collections.Generic;
+
+namespace PersonalLogManager.Service.TextBuilding
+{
+    public abstract class PersonalLogTextBuilderBase
+    {
+        public string GetDiscriminator(Dictionary<string, string> data)
+        {
+            data.TryGetValue("discriminator", out string discriminator);
+
+            if (string.IsNullOrWhiteSpace(discriminator))
+            {
+                data.TryGetValue("account", out discriminator);
+            }
+
+            if (string.IsNullOrWhiteSpace(discriminator))
+            {
+                data.TryGetValue("account_id", out discriminator);
+            }
+
+            if (string.IsNullOrWhiteSpace(discriminator))
+            {
+                data.TryGetValue("username", out discriminator);
+            }
+
+            if (string.IsNullOrWhiteSpace(discriminator))
+            {
+                data.TryGetValue("phone_number", out discriminator);
+            }
+
+            if (string.IsNullOrWhiteSpace(discriminator))
+            {
+                data.TryGetValue("email_address", out discriminator);
+            }
+
+            return discriminator;
+        }
+    }
+}
