+++ ServiceCollectionExtensions.cs
@@ -33,7 +33,6 @@ namespace PersonalLogManager
         public static IServiceCollection AddCustomServices(this IServiceCollection services) => services
             .AddSingleton<IFileRepository<PersonalLogEntity>>(x => new JsonRepository<PersonalLogEntity>(dataStoreSettings.TextLogStorePath))
             .AddSingleton<IPersonalLogTextBuilderFactory, PersonalLogTextBuilderFactory>()
-            .AddSingleton<IPersonalLogTextBuilder, PersonalLogTextBuilder>()
             .AddSingleton<IPersonalLogService, PersonalLogService>()
             .AddAutoMapper(typeof(DataAccessMappingProfile))
             .AddAutoMapper(typeof(ServiceMappingProfile));
